+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/mining/Durium.java
@@ -5,7 +5,7 @@ import insane96mcp.iguanatweaksexpanded.item.ITEArmorMaterial;
 import insane96mcp.iguanatweaksexpanded.module.Modules;
 import insane96mcp.iguanatweaksexpanded.module.misc.ITEDataPacks;
 import insane96mcp.iguanatweaksexpanded.setup.ITERegistries;
-import insane96mcp.iguanatweaksexpanded.setup.IntegratedDataPack;
+import insane96mcp.iguanatweaksexpanded.setup.IntegratedPack;
 import insane96mcp.iguanatweaksexpanded.setup.registry.SimpleBlockWithItem;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
@@ -78,6 +78,6 @@ Durium extends Feature {
 
 	public Durium(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
-		IntegratedDataPack.INTEGRATED_DATA_PACKS.add(new IntegratedDataPack(PackType.SERVER_DATA, "durium_lodestone", Component.literal("IguanaTweaks Expanded Durium Lodestone"), () -> this.isEnabled() && !ITEDataPacks.disableAllDataPacks && duriumLodestone));
+		IntegratedPack.addPack(new IntegratedPack(PackType.SERVER_DATA, "durium_lodestone", Component.literal("IguanaTweaks Expanded Durium Lodestone"), () -> this.isEnabled() && !ITEDataPacks.disableAllDataPacks && duriumLodestone));
 	}
 }