+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/mining/multiblockfurnaces/MultiBlockFurnaces.java
@@ -15,7 +15,7 @@ import insane96mcp.iguanatweaksexpanded.module.mining.multiblockfurnaces.invento
 import insane96mcp.iguanatweaksexpanded.module.mining.multiblockfurnaces.inventory.MultiBlockSoulBlastFurnaceMenu;
 import insane96mcp.iguanatweaksexpanded.module.misc.ITEDataPacks;
 import insane96mcp.iguanatweaksexpanded.setup.ITERegistries;
-import insane96mcp.iguanatweaksexpanded.setup.IntegratedDataPack;
+import insane96mcp.iguanatweaksexpanded.setup.IntegratedPack;
 import insane96mcp.iguanatweaksexpanded.setup.registry.SimpleBlockWithItem;
 import insane96mcp.iguanatweaksexpanded.utils.ClientUtils;
 import insane96mcp.insanelib.base.Feature;
@@ -81,7 +81,7 @@ public class MultiBlockFurnaces extends Feature {
 
 	public MultiBlockFurnaces(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
-		IntegratedDataPack.INTEGRATED_DATA_PACKS.add(new IntegratedDataPack(PackType.SERVER_DATA, "multi_block_blast_furnace", Component.literal("IguanaTweaks Expanded Multi Block Blast Furnace"), () -> this.isEnabled() && !ITEDataPacks.disableAllDataPacks && blastFurnaceDataPack));
+		IntegratedPack.addPack(new IntegratedPack(PackType.SERVER_DATA, "multi_block_blast_furnace", Component.literal("IguanaTweaks Expanded Multi Block Blast Furnace"), () -> this.isEnabled() && !ITEDataPacks.disableAllDataPacks && blastFurnaceDataPack));
 	}
 
 	@SubscribeEvent